refactor: 이메일 관련 도메인 객체들 패키지 경로를 User로 변경
