+++ src/main/java/koreatech/in/domain/User/LocalParts.java
@@ -1,4 +1,4 @@
-package koreatech.in.domain.User.owner;
+package koreatech.in.domain.User;
 
 import lombok.Getter;
