+++ src/main/java/koreatech/in/domain/User/owner/OwnerInCertification.java
@@ -1,5 +1,6 @@
 package koreatech.in.domain.User.owner;
 
+import koreatech.in.domain.User.EmailAddress;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
