+++ src/main/java/koreatech/in/mapstruct/OwnerConverter.java
@@ -3,9 +3,9 @@ package koreatech.in.mapstruct;
 import java.util.List;
 import java.util.stream.Collectors;
 import koreatech.in.domain.User.User;
-import koreatech.in.domain.User.owner.Domain;
-import koreatech.in.domain.User.owner.EmailAddress;
-import koreatech.in.domain.User.owner.LocalParts;
+import koreatech.in.domain.User.Domain;
+import koreatech.in.domain.User.EmailAddress;
+import koreatech.in.domain.User.LocalParts;
 import koreatech.in.domain.User.owner.Owner;
 import koreatech.in.domain.User.owner.OwnerInCertification;
 import koreatech.in.domain.User.owner.OwnerShopAttachment;
